removed more specialized components and made the mechanism components a bit more generic
