--- src/components/Mechanism/AerosolSpeciesTab.js
@@ -1,12 +1,0 @@
-import React from "react";
-import SpeciesTab from "./SpeciesTab.js";
-
-function AerosolSpeciesTab() {
-  return (
-    <SpeciesTab
-      type="aerosol"
-    />
-  );
-}
-
-export default AerosolSpeciesTab;
