--- src/components/Mechanism/GasSpeciesTab.js
@@ -1,12 +1,0 @@
-import React from "react";
-import SpeciesTab from "./SpeciesTab.js";
-
-function GasSpeciesTab() {
-  return (
-    <SpeciesTab
-      type="gas"
-    />
-  );
-}
-
-export default GasSpeciesTab;
