+++ src/components/Mechanism/MechanismTab.js
@@ -3,25 +3,41 @@ import { connect } from "react-redux";
 import { Container, Row, Col } from "react-bootstrap";
 import { Link } from "gatsby";
 
-import SpeciesList from "./SpeciesList";
-import SpeciesDetail from "./SpeciesDetail.js";
+import List from "./List";
+import Detail from "./Detail";
 
-function SpeciesTab({ type }) {
-  const [details, setDetails] = useState({});
-
-  return (
-    <>
+const InstructionsComponent = ({ type }) => {
+  if (type === "reactions") {
+    return (
+      <p className="lead-muted p-2">
+        Select a reaction from the list to view/edit its properties, or add a new reaction to the mechanism.
+        The chemical species available to participate in reactions can be modified{" "}
+        <Link to="mechanism/species">here</Link>.
+      </p>
+    )
+  }
+  else{
+    return (
       <p className="lead-muted p-2">
         Select a chemical species from the list to view/edit its properties, or add a new chemical species to the mechansim.
         The chemical species you add here will be available to participate in <Link to="mechanism/reactions">reactions</Link> 
         &nbsp;and can be include in the <Link to="conditions">model conditions</Link>.
       </p>
-      <Container fluid className="p-2 d-flex flex-column overflow-hidden">
+    )
+  }
+}
+
+function MechanismTab({ type, }) {
+  const [details, setDetails] = useState({});
+  return (
+    <>
+      <InstructionsComponent type={type} />
+      <Container fluid className="p-2 d-flex flex-column vh-100 overflow-hidden">
         <Row className="flex-grow-1 overflow-hidden">
           <Col md={4} lg={4} className="mh-100 overflow-auto">
             <Row className="flex-shrink-0">
               <Col>
-                <SpeciesList
+                <List
                   type={type}
                   details={details}
                   setDetails={setDetails}
@@ -33,9 +49,9 @@ function SpeciesTab({ type }) {
             <Row className="flex-shrink-0">
               <Col className="species-detail">
                 {Object.keys(details).map(key => (
-                  <SpeciesDetail
+                  <Detail
                     type={type}
-                    species={details[key]}
+                    item={details[key]}
                     key={key}
                     details={details}
                     setDetails={setDetails}
@@ -50,4 +66,4 @@ function SpeciesTab({ type }) {
   );
 }
 
-export default connect()(SpeciesTab);
+export default connect()(MechanismTab);
