+++ src/components/Mechanism/Reactions/AddProduct.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { connect } from "react-redux";
-import { addProduct } from "../../redux/actions";
+import { addProduct } from "../../../redux/actions";
 
 const AddProduct = (props) => {
 
@@ -14,9 +14,9 @@ const AddProduct = (props) => {
 
   return (
     <button type="button"
-            className="btn btn-primary add-element"
-            onClick={() => handleAddProduct({ name: undefined, yield: 1.0 })}
-            onKeyDown={() => handleAddProduct({ name: undefined, yield: 1.0 })}>
+      className="btn btn-primary add-element"
+      onClick={() => handleAddProduct({ name: undefined, yield: 1.0 })}
+      onKeyDown={() => handleAddProduct({ name: undefined, yield: 1.0 })}>
       <span className="oi oi-plus" toggle="tooltip" aria-hidden="true" title="Add product"></span>
     </button>
   );
