+++ src/components/Mechanism/Reactions/AddReactant.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { connect } from "react-redux";
-import { addReactant } from "../../redux/actions";
+import { addReactant } from "../../../redux/actions";
 
 const AddReactant = (props) => {
 
@@ -10,9 +10,9 @@ const AddReactant = (props) => {
 
   return (
     <button type="button"
-            className="btn btn-primary add-element"
-            onClick={() => handleAddReactant({ name: undefined, qty: 1 })}
-            onKeyDown={() => handleAddReactant({ name: undefined, qty: 1 })}>
+      className="btn btn-primary add-element"
+      onClick={() => handleAddReactant({ name: undefined, qty: 1 })}
+      onKeyDown={() => handleAddReactant({ name: undefined, qty: 1 })}>
       <span className="oi oi-plus" toggle="tooltip" aria-hidden="true" title="Add reactant"></span>
     </button>
   );
