+++ src/components/Mechanism/Reactions/ProductList.js
@@ -2,7 +2,7 @@ import React from "react";
 import { connect } from "react-redux";
 import AddProduct from "./AddProduct";
 import Product from "./Product";
-import { getProducts } from "../../redux/selectors";
+import { getProducts } from "../../../redux/selectors";
 
 const ProductList = props => {
   return (
@@ -11,7 +11,7 @@ const ProductList = props => {
         <div className="card-header d-flex justify-content-between">
           <h3 className="my-0 fw-normal">{props.schema.label}</h3>
           <AddProduct reactionId={props.reactionId}
-                      schema={props.schema} />
+            schema={props.schema} />
         </div>
         <div className="card-body">
           <div className="form-group array-elements array-elements-products container-fluid">
@@ -19,15 +19,15 @@ const ProductList = props => {
               props.products.map((product, index) => {
                 return (
                   <div key={`product-${product.name}`}
-                       className={`row flex-nowrap array-element array-element-${index}`}
-                       array-element-index={index}>
+                    className={`row flex-nowrap array-element array-element-${index}`}
+                    array-element-index={index}>
                     <Product reactionId={props.reactionId}
-                             product={product}
-                             schema={props.schema} />
+                      product={product}
+                      schema={props.schema} />
                   </div>
                 );
               })
-              : null }
+              : null}
           </div>
         </div>
       </div>
