+++ src/components/Mechanism/Species/PropertyList.js
@@ -1,10 +1,35 @@
 import React, { useEffect, useState } from 'react';
 import { connect } from "react-redux";
-import Property from "./Property";
 import AddProperty from "./AddProperty";
-import { aerosol_options, gas_options } from "./shared_properties"
-import { getMechanism } from "../../redux/selectors";
-import { addProperty, addAerosolProperty } from "../../redux/actions";
+import { aerosol_options, gas_options } from "../shared_properties"
+import { getMechanism } from "../../../redux/selectors";
+import { addProperty, addAerosolProperty } from "../../../redux/actions";
+
+const Property = ({ property, addAction, speciesName }) => {
+  const handleAddProperty = (event) => {
+    addAction({
+      property: {
+        ...property,
+        value: event.target.value
+      },
+      speciesName: speciesName
+    });
+  };
+
+  return (
+    <div className="input-group mb-3" property={property.name} data-type={property.data_type}>
+      <div className="input-group-prepend">
+        <span className="input-group-text">{property.name}</span>
+      </div>
+      <input type="text"
+        className="form-control"
+        placeholder="Property value"
+        value={property.value}
+        onChange={handleAddProperty}>
+      </input>
+    </div>
+  );
+};
 
 function PropertyList({ options, properties, speciesName, type, addAction }) {
   const [localProperties, setLocalProperties] = useState([]);
@@ -21,15 +46,15 @@ function PropertyList({ options, properties, speciesName, type, addAction }) {
       {
         localProperties?.map(property => {
           return <Property key={property.name}
-                           property={property}
-                           speciesName={speciesName}
-                           addAction={addAction}
-                           />;
+            property={property}
+            speciesName={speciesName}
+            addAction={addAction}
+          />;
         })
       }
-      <AddProperty type={type} speciesName={speciesName} options={options} addAction={addAction}/>
+      <AddProperty type={type} speciesName={speciesName} options={options} addAction={addAction} />
       <p>
-      You may specify any property you like, but this is only <em>necessary</em> under certain circumstances (i.e., when the species participates in a reaction that requires the property be set). You will be prompted to set the property when it is required.
+        You may specify any property you like, but this is only <em>necessary</em> under certain circumstances (i.e., when the species participates in a reaction that requires the property be set). You will be prompted to set the property when it is required.
       </p>
     </div>
   );
@@ -43,7 +68,7 @@ const mapStateToProps = (state, ownProps) => {
   const mechanism = getMechanism(state);
   const speciesList = type === "gas" ? mechanism.gasSpecies : type === "aerosol" ? mechanism.aerosolSpecies : [];
 
-  const species = speciesList?.find(spec => spec.name === speciesName );
+  const species = speciesList?.find(spec => spec.name === speciesName);
 
   return {
     properties: species ? species.properties : [],
