+++ src/components/Mechanism/index.js
@@ -1,3 +1,1 @@
-export { default as AerosolSpeciesTab } from './AerosolSpeciesTab'
-export { default as GasSpeciesTab } from './GasSpeciesTab'
-export { default as ReactionsTab } from './ReactionsTab'
+export { default as MechanismTab } from './MechanismTab'
